implemented recylerview and toolbar
