+++ app/build.gradle
@@ -51,4 +51,5 @@ dependencies {
 
 
     implementation 'androidx.activity:activity-ktx:1.4.0'
+
 }
