+++ app/src/main/AndroidManifest.xml
@@ -11,19 +11,19 @@
         android:theme="@style/Theme.PMPDPlayGround">
 
         <activity
-            android:name="com.tijanidian.mvvm.presentation.MvvmActivity"
+            android:name="com.tijanidian.pmpd_playground.mvvm.presentation.MvvmActivity"
             android:exported="true" />
         <activity
-            android:name=".MainActivity"
+            android:name=".pruebamvvm.MainActivity"
             android:exported="true" />
         <activity
-            android:name="ut2.UiActivity"
+            android:name="com.tijanidian.pmpd_playground.pruebamvvm.UiActivity"
             android:exported="true" />
         <activity
-            android:name="ut2.UiTestGlovoActivity"
+            android:name="com.tijanidian.pmpd_playground.pruebamvvm.UiTestGlovoActivity"
             android:exported="true"/>
             <activity
-                android:name="com.tijanidian.ut02ex06form.presentation.Ut02ex06Activity"
+                android:name="com.tijanidian.pmpd_playground.ut2.ut02ex06form.presentation.Ut02ex06Activity"
                 android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
