--- app/src/main/java/com/tijanidian/mvvm/data/MvvmDataRepository.kt
@@ -1,22 +1,0 @@
-package com.tijanidian.mvvm.data
-
-import com.tijanidian.mvvm.domain.MvvmModel
-import com.tijanidian.mvvm.domain.MvvmRepository
-
-class MvvmDataRepository: MvvmRepository {
-    override fun getMvvm(): MvvmModel {
-        TODO("Not yet implemented")
-    }
-
-    override fun saveMvvm(mvvmModel: MvvmModel) {
-        TODO("Not yet implemented")
-    }
-
-    override fun removeMvvm(id: String) {
-        TODO("Not yet implemented")
-    }
-
-    override fun getMvvmAll(): List<MvvmModel> {
-        TODO("Not yet implemented")
-    }
-}
