--- app/src/main/java/com/tijanidian/mvvm/domain/GetMvvmUseCase.kt
@@ -1,9 +1,0 @@
-package com.tijanidian.mvvm.domain
-
-class GetMvvmUseCase(val mvvmRepository: MvvmRepository){
-
-    fun execute(paramsIn:String):MvvmModel{
-        val model=mvvmRepository.getMvvm()
-        return model
-    }
-}
