--- app/src/main/java/com/tijanidian/mvvm/domain/MvvmRepository.kt
@@ -1,8 +1,0 @@
-package com.tijanidian.mvvm.domain
-
-interface MvvmRepository {
-    fun getMvvm():MvvmModel
-    fun saveMvvm(mvvmModel: MvvmModel)
-    fun removeMvvm(id:String)
-    fun getMvvmAll():List<MvvmModel>
-}
