--- app/src/main/java/com/tijanidian/mvvm/presentation/MvvmViewModel.kt
@@ -1,9 +1,0 @@
-package com.tijanidian.mvvm.presentation
-
-import androidx.lifecycle.ViewModel
-import com.tijanidian.mvvm.domain.GetMvvmUseCase
-
-class MvvmViewModel(private val getMvvmUseCase:GetMvvmUseCase): ViewModel() {
-
-    fun getAllMvvm()=getMvvmUseCase.execute("1")
-}
