+++ app/src/main/java/com/tijanidian/pmpd_playground/commons/serializer/JsonSerializer.kt
@@ -1,0 +1,6 @@
+package com.tijanidian.pmpd_playground.commons.serializer
+
+interface JsonSerializer {
+    fun <T> toJson(obj: T, typeClass: Class<T>): String
+    fun <T> fromJson(json: String, typeClass: Class<T>): T
+}
