+++ app/src/main/java/com/tijanidian/pmpd_playground/mvvm/data/MvvmDataRepository.kt
@@ -1,0 +1,22 @@
+package com.tijanidian.pmpd_playground.mvvm.data
+
+import com.tijanidian.pmpd_playground.mvvm.domain.MvvmModel
+import com.tijanidian.pmpd_playground.mvvm.domain.MvvmRepository
+
+class MvvmDataRepository: MvvmRepository {
+    override fun getMvvm(): MvvmModel {
+        TODO("Not yet implemented")
+    }
+
+    override fun saveMvvm(mvvmModel: MvvmModel) {
+        TODO("Not yet implemented")
+    }
+
+    override fun removeMvvm(id: String) {
+        TODO("Not yet implemented")
+    }
+
+    override fun getMvvmAll(): List<MvvmModel> {
+        TODO("Not yet implemented")
+    }
+}
