+++ app/src/main/java/com/tijanidian/pmpd_playground/mvvm/domain/MvvmRepository.kt
@@ -1,0 +1,8 @@
+package com.tijanidian.pmpd_playground.mvvm.domain
+
+interface MvvmRepository {
+    fun getMvvm():MvvmModel
+    fun saveMvvm(mvvmModel: MvvmModel)
+    fun removeMvvm(id:String)
+    fun getMvvmAll():List<MvvmModel>
+}
