+++ app/src/main/java/com/tijanidian/pmpd_playground/mvvm/presentation/MvvmActivity.kt
@@ -1,0 +1,20 @@
+package com.tijanidian.pmpd_playground.mvvm.presentation
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.activity.viewModels
+import com.tijanidian.pmpd_playground.R
+
+class MvvmActivity : AppCompatActivity() {
+    private val viewMode:MvvmViewModel by viewModels()
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_mvvm)
+        render()
+    }
+
+    private fun render(){
+        val mvvmModel=viewMode.getAllMvvm()
+
+    }
+}
