+++ app/src/main/java/com/tijanidian/pmpd_playground/pruebamvvm/GlovoModel.kt
@@ -1,0 +1,9 @@
+package com.tijanidian.pmpd_playground.pruebamvvm
+
+import com.tijanidian.pmpd_playground.R
+
+data class GlovoModel(val principal: Int = R.string.glovo_principal_text,
+                      val segundario:Int=R.string.glovo_secundario_text,
+                      val leftIcon:Int=R.string.label_price,
+                      val rightIcon:Int=R.string.text_right_icon
+                      )
