+++ app/src/main/java/com/tijanidian/pmpd_playground/pruebamvvm/UiTestGlovoViewModel.kt
@@ -1,0 +1,5 @@
+package com.tijanidian.pmpd_playground.pruebamvvm
+
+class UiTestGlovoViewModel {
+    fun getGlovoModel() = GlovoModel()
+}
