+++ app/src/main/java/com/tijanidian/pmpd_playground/ut2/ut02ex06form/domain/Models.kt
@@ -1,0 +1,14 @@
+package com.tijanidian.pmpd_playground.ut2.ut02ex06form.domain
+
+
+class PlayerModel (id:String,name:String) {
+}
+
+data class PlayerModelFootball(
+    val idPlayer:Int,
+    val name:String,
+    val community:String,
+    val sex:String,
+    val firtsPosition:String,
+    val secondPosition:String,
+)
