+++ app/src/main/java/com/tijanidian/pmpd_playground/ut2/ut02ex06form/domain/SavePlayerUseCase.kt
@@ -1,0 +1,10 @@
+package com.tijanidian.pmpd_playground.ut2.ut02ex06form.domain
+
+class SavePlayerUseCase {
+
+    fun execute(param: Param){
+
+    }
+    data class Param(val name:String,val surname:String)
+
+}
