+++ app/src/main/java/com/tijanidian/pmpd_playground/ut2/ut02ex06form/presentation/viewmodel/FromFragmentViewModel.kt
@@ -1,0 +1,4 @@
+package com.tijanidian.pmpd_playground.ut2.ut02ex06form.presentation.viewmodel
+
+class FromFragmentViewModel {
+}
