--- app/src/main/java/com/tijanidian/ut02ex06form/domain/Models.kt
@@ -1,13 +1,0 @@
-package com.tijanidian.ut02ex06form.domain
-
-
-class PlayerModel (id:String,name:String) {
-}
-
-data class PlayerModelFootball(
-    val name:String,
-    val community:String,
-    val sex:String,
-    val firtsPosition:String,
-    val secondPosition:String,
-)
