--- app/src/main/java/com/tijanidian/ut02ex06form/domain/SavePlayerUseCase.kt
@@ -1,10 +1,0 @@
-package com.tijanidian.ut02ex06form.domain
-
-class SavePlayerUseCase {
-
-    fun execute(param: Param){
-
-    }
-    data class Param(val name:String,val surname:String)
-
-}
