--- app/src/main/java/ut2/UiTestGlovoViewModel.kt
@@ -1,5 +1,0 @@
-package ut2
-
-class UiTestGlovoViewModel {
-    fun getGlovoModel() = GlovoModel()
-}
