+++ app/src/main/res/layout/activity_mvvm.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.tijanidian.mvvm.presentation.MvvmActivity">
+    tools:context=".mvvm.presentation.MvvmActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
