+++ app/src/main/res/layout/activity_ut02ex06.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.tijanidian.ut02ex06form.presentation.Ut02ex06Activity">
+    tools:context="com.tijanidian.pmpd_playground.ut2.ut02ex06form.presentation.Ut02ex06Activity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/main_bar"
