+++ app/src/main/res/values/strings.xml
@@ -33,4 +33,9 @@
 
 
 
+    <!--Form players-->
+
+    <string name="form_title">Formulario</string>
+    <string name="list_title">Lista</string>
+
 </resources>
